face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
insightface==0.7.3
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
numpy>=1.24,<2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
